+++ src/main/java/com/wayqui/transaq/dto/TransactionStatusDto.java
@@ -4,12 +4,14 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 @Setter
 public class TransactionStatusDto {
     private String reference;
     private TransactionStatus status;
-    private Double amount;
-    private Double fee;
+    private BigDecimal amount;
+    private BigDecimal fee;
 }
